remove native watcher
