+++ internal/langserver/handlers/service.go
@@ -49,8 +49,6 @@ type service struct {
 	fs            *filesystem.Filesystem
 	modStore      *state.ModuleStore
 	schemaStore   *state.ProviderSchemaStore
-	watcher       module.Watcher
-	newWatcher    module.WatcherFactory
 	tfDiscoFunc   discovery.DiscoveryFunc
 	tfExecFactory exec.ExecutorFactory
 	tfExecOpts    *exec.ExecutorOpts
@@ -79,7 +77,6 @@ func NewSession(srvCtx context.Context) session.Session {
 		srvCtx:        srvCtx,
 		sessCtx:       sessCtx,
 		stopSession:   stopSession,
-		newWatcher:    module.NewWatcher,
 		tfDiscoFunc:   d.LookPath,
 		tfExecFactory: exec.NewExecutor,
 		telemetry:     &telemetry.NoopSender{},
@@ -153,7 +150,6 @@ func (svc *service) Assigner() (jrpc2.Assigner, error) {
 			if err != nil {
 				return nil, err
 			}
-			ctx = lsctx.WithWatcher(ctx, svc.watcher)
 			return handle(ctx, req, svc.TextDocumentDidOpen)
 		},
 		"textDocument/didClose": func(ctx context.Context, req *jrpc2.Request) (interface{}, error) {
@@ -288,8 +284,6 @@ func (svc *service) Assigner() (jrpc2.Assigner, error) {
 				return nil, err
 			}
 
-			ctx = lsctx.WithWatcher(ctx, svc.watcher)
-
 			return handle(ctx, req, svc.DidChangeWorkspaceFolders)
 		},
 		"workspace/didChangeWatchedFiles": func(ctx context.Context, req *jrpc2.Request) (interface{}, error) {
@@ -315,7 +309,6 @@ func (svc *service) Assigner() (jrpc2.Assigner, error) {
 			}
 
 			ctx = lsctx.WithCommandPrefix(ctx, &commandPrefix)
-			ctx = lsctx.WithWatcher(ctx, svc.watcher)
 			ctx = lsctx.WithRootDirectory(ctx, &rootDir)
 			ctx = lsctx.WithDiagnosticsNotifier(ctx, svc.diagsNotifier)
 			ctx = ilsp.ContextWithClientName(ctx, &clientName)
@@ -501,17 +494,6 @@ func (svc *service) configureSessionDependencies(ctx context.Context, cfgOpts *s
 	svc.closedDirWalker.Collector = svc.walkerCollector
 	svc.openDirWalker.SetLogger(svc.logger)
 
-	ww, err := svc.newWatcher(svc.fs, svc.modStore, svc.stateStore.ProviderSchemas, svc.stateStore.JobStore, svc.tfExecFactory)
-	if err != nil {
-		return err
-	}
-	svc.watcher = ww
-	svc.watcher.SetLogger(svc.logger)
-	err = svc.watcher.Start(ctx)
-	if err != nil {
-		return err
-	}
-
 	return nil
 }
 
@@ -544,16 +526,6 @@ func (svc *service) shutdown() {
 		svc.logger.Printf("stopping openDirWalker for session ...")
 		svc.openDirWalker.Stop()
 		svc.logger.Printf("openDirWalker stopped")
-	}
-
-	if svc.watcher != nil {
-		svc.logger.Println("stopping watcher for session ...")
-		err := svc.watcher.Stop()
-		if err != nil {
-			svc.logger.Println("unable to stop watcher for session:", err)
-		} else {
-			svc.logger.Println("watcher stopped")
-		}
 	}
 
 	if svc.closedDirIndexer != nil {
